@@ -1,193 +1,181 @@
 {
-  "name": "angular-starter",
-  "version": "6.0.0",
-  "description": "An Angular Webpack Starter kit featuring Angular (Router, Http, Forms, Services, Tests, E2E, Coverage), Karma, Protractor, Jasmine, Istanbul, TypeScript, and Webpack by AngularClass",
-  "keywords": [
-    "angular",
-    "angular2",
-    "angular4",
-    "webpack",
-    "typescript"
-  ],
-  "author": "Patrick Stapleton <patrick@angularclass.com>",
-  "homepage": "https://github.com/AngularClass/angular-starter",
-  "license": "MIT",
-  "scripts": {
-    "build:aot:prod": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
-    "build:aot": "npm run build:aot:prod",
-    "build:dev": "npm run clean:dist && npm run webpack -- --config config/webpack.dev.js --progress --profile",
-    "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
-    "build:prod": "npm run clean:dist && npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
-    "build": "npm run build:dev",
-    "ci:aot": "npm run lint && npm run test && npm run build:aot && npm run e2e",
-    "ci:jit": "npm run lint && npm run test && npm run build:prod && npm run e2e",
-    "ci:nobuild": "npm run lint && npm test && npm run e2e",
-    "ci:testall": "npm run lint && npm run test && npm run build:prod && npm run e2e && npm run build:aot && npm run e2e",
-    "ci:travis": "npm run lint && npm run test && npm run build:aot && npm run e2e:travis",
-    "ci": "npm run ci:testall",
-    "clean:dll": "npm run rimraf -- dll",
-    "clean:aot": "npm run rimraf -- compiled",
-    "clean:dist": "npm run rimraf -- dist",
-    "clean:install": "npm set progress=false && npm install",
-    "clean": "npm cache clean --force && npm run rimraf -- node_modules doc coverage dist compiled dll",
-    "docker": "docker",
-    "docs": "npm run typedoc -- --options typedoc.json --exclude '**/*.spec.ts' ./src/",
-    "e2e:live": "npm-run-all -p -r server:prod:ci protractor:live",
-    "e2e:travis": "npm-run-all -p -r server:prod:ci protractor:delay",
-    "e2e": "npm-run-all -p -r server:prod:ci protractor",
-    "github-deploy:dev": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubDev",
-    "github-deploy:prod": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubProd",
-    "github-deploy": "npm run github-deploy:dev",
-    "lint": "npm run tslint \"src/**/*.ts\"",
-    "node": "node",
-    "postinstall": "npm run webdriver:update",
-    "postversion": "git push && git push --tags",
-    "preclean:install": "npm run clean",
-    "preversion": "npm test",
-    "protractor": "protractor",
-    "protractor:delay": "sleep 3 && npm run protractor",
-    "protractor:live": "protractor --elementExplorer",
-    "rimraf": "rimraf",
-    "server:dev:hmr": "npm run server:dev -- --inline --hot",
-    "server:dev": "npm run webpack-dev-server -- --config config/webpack.dev.js --open --progress --profile --watch --content-base src/",
-    "server:prod": "http-server dist -c-1 --cors",
-    "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
-    "server": "npm run server:dev",
-    "start:hmr": "npm run server:dev:hmr",
-    "start": "npm run server:dev",
-    "test": "npm run lint && karma start",
-    "tslint": "tslint",
-    "typedoc": "typedoc",
-    "version": "npm run build",
-    "watch:dev:hmr": "npm run watch:dev -- --hot",
-    "watch:dev": "npm run build:dev -- --watch",
-    "watch:prod": "npm run build:prod -- --watch",
-    "watch:test": "npm run test -- --auto-watch --no-single-run",
-    "watch": "npm run watch:dev",
-    "webdriver-manager": "webdriver-manager",
-    "webdriver:start": "npm run webdriver-manager start",
-    "webdriver:update": "webdriver-manager update",
-    "webpack-dev-server": "node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-    "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js",
-    "precommit": "lint-staged"
-  },
-  "lint-staged": {
-    "src/**/*.ts": [
-      "prettier --write",
-      "git add"
+    "name": "angular-starter",
+    "version": "7.0.0",
+    "description": "An Angular Webpack Starter kit featuring Angular (Router, Http, Forms, Services, Tests, E2E, Coverage), Karma, Protractor, Jasmine, Istanbul, TypeScript, and Webpack by AngularClass",
+    "keywords": [
+      "angular",
+      "angular2",
+      "angular4",
+      "angular5",
+      "webpack",
+      "typescript",
+      "tipe",
+      "tipe.io"
     ],
-    "src/assets/**/*.json": [
-      "prettier --write",
-      "git add"
-    ],
-    "src/**/*.scss": [
-      "prettier --write",
-      "git add"
-    ]
-  },
-  "dependencies": {
-    "@angular/animations": "^5.0.4",
-    "@angular/common": "^5.0.4",
-    "@angular/compiler": "^5.0.4",
-    "@angular/core": "^5.0.4",
-    "@angular/forms": "^5.0.4",
-    "@angular/http": "^5.0.4",
-    "@angular/platform-browser": "^5.0.4",
-    "@angular/platform-browser-dynamic": "^5.0.4",
-    "@angular/platform-server": "^5.0.4",
-    "@angular/router": "^5.0.4",
-    "@angularclass/hmr": "~1.2.2",
-    "@angularclass/hmr-loader": "~3.0.2",
-    "core-js": "^2.4.1",
-    "http-server": "^0.9.0",
-    "ie-shim": "^0.1.0",
-    "install": "^0.10.2",
-    "reflect-metadata": "^0.1.10",
-    "rxjs": "~5.5.2",
-    "zone.js": "^0.8.18"
-  },
-  "devDependencies": {
-    "@angular/compiler-cli": "^5.0.4",
-    "@ngtools/webpack": "^1.8.5",
-    "@types/hammerjs": "^2.0.34",
-    "@types/jasmine": "2.5.45",
-    "@types/node": "^7.0.13",
-    "@types/source-map": "^0.5.0",
-    "@types/uglify-js": "^2.6.28",
-    "@types/webpack": "^2.2.15",
-    "add-asset-html-webpack-plugin": "^1.0.2",
-    "angular2-template-loader": "^0.6.2",
-    "assets-webpack-plugin": "^3.5.1",
-    "autoprefixer": "^7.1.2",
-    "awesome-typescript-loader": "~3.1.2",
-    "codelyzer": "~3.2.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.0.0",
-    "css-loader": "^0.28.0",
-    "cssnano": "^3.10.0",
-    "cssnext": "^1.8.4",
-    "exports-loader": "^0.6.4",
-    "expose-loader": "^0.7.3",
-    "extract-text-webpack-plugin": "~2.1.0",
-    "file-loader": "^0.11.1",
-    "find-root": "^1.0.0",
-    "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "^2.28.0",
-    "imports-loader": "^0.7.1",
-    "inline-manifest-webpack-plugin": "^3.0.1",
-    "istanbul-instrumenter-loader": "2.0.0",
-    "jasmine-core": "^2.5.2",
-    "json-loader": "^0.5.4",
-    "karma": "^1.6.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-mocha-reporter": "^2.2.3",
-    "karma-remap-coverage": "^0.1.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "~2.0.3",
-    "lint-staged": "^4.2.3",
-    "ng-router-loader": "^2.1.0",
-    "ngc-webpack": "^4.1.2",
-    "node-sass": "^4.5.2",
-    "npm-run-all": "^4.0.2",
-    "optimize-js-plugin": "0.0.4",
-    "parse5": "^3.0.2",
-    "postcss-cssnext": "^3.0.2",
-    "postcss-import": "^10.0.0",
-    "postcss-loader": "^2.0.6",
-    "preload-webpack-plugin": "^1.2.2",
-    "protractor": "^5.1.1",
-    "raw-loader": "0.5.1",
-    "rimraf": "~2.6.1",
-    "sass-loader": "^6.0.3",
-    "script-ext-html-webpack-plugin": "^1.7.1",
-    "source-map-loader": "^0.2.1",
-    "string-replace-loader": "~1.2.0",
-    "style-loader": "^0.18.2",
-    "to-string-loader": "^1.1.5",
-    "ts-node": "^3.0.2",
-    "tslib": "^1.6.1",
-    "tslint": "~5.7.0",
-    "tslint-loader": "^3.5.2",
-    "typedoc": "^0.7.1",
-    "typescript": "^2.0.2",
-    "url-loader": "^0.5.8",
-    "webpack": "^3.9.1",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-dev-server": "^2.9.5",
-    "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
-    "webpack-merge": "~4.1.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/AngularClass/angular-starter.git"
-  },
-  "bugs": {
-    "url": "https://github.com/AngularClass/angular-starter/issues"
-  },
-  "engines": {
-    "node": ">= 4.2.1",
-    "npm": ">= 3"
+    "author": "Patrick Stapleton <patrick@tipe.io>",
+    "homepage": "https://github.com/gdi2290/angular-starter",
+    "license": "MIT",
+    "scripts": {
+      "build:aot:prod": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
+      "build:aot": "npm run build:aot:prod",
+      "build:aot:dev": "cross-env BUILD_AOT=1 npm run build:dev",
+      "build:dev": "npm run clean:dist && npm run webpack -- --config config/webpack.dev.js --progress --profile",
+      "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
+      "build:prod": "npm run clean:dist && npm run webpack -- --config config/webpack.prod.js  --progress --profile --bail",
+      "build": "npm run build:dev",
+      "ci:aot": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:aot && npm run e2e",
+      "ci:jit": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:prod && npm run e2e",
+      "ci:nobuild": "npm run lint && npm test && npm run e2e",
+      "ci:testall": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:prod && npm run e2e && npm run build:aot && npm run e2e",
+      "ci:travis": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:aot && npm run e2e:travis",
+      "ci": "npm run ci:testall",
+      "clean:dll": "npm run rimraf -- dll",
+      "clean:aot": "npm run rimraf -- compiled",
+      "clean:dist": "npm run rimraf -- dist",
+      "clean:install": "npm set progress=false && npm install",
+      "clean": "npm cache clean --force && npm run rimraf -- node_modules doc coverage dist compiled dll",
+      "docker": "docker",
+      "docs": "npm run typedoc -- --options typedoc.json --exclude '**/*.spec.ts' ./src/",
+      "docs:compodoc": "compodoc -p tsconfig.json",
+      "docs:compodoc:serve": "compodoc -p tsconfig.json -s",
+      "docs:compodoc:serve:watch": "compodoc -p tsconfig.json -s -w",
+      "e2e:live": "npm-run-all -p -r server:prod:ci protractor:live",
+      "e2e:travis": "npm-run-all -p -r server:prod:ci protractor:delay",
+      "e2e": "npm-run-all -p -r server:prod:ci protractor",
+      "github-deploy:dev": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubDev",
+      "github-deploy:prod": "npm run webpack -- --config config/webpack.github-deploy.js --progress --profile --env.githubProd",
+      "github-deploy": "npm run github-deploy:dev",
+      "lint": "npm run tslint \"src/**/*.ts\"",
+      "node": "node",
+      "postinstall": "npm run webdriver:update",
+      "postversion": "git push && git push --tags",
+      "preclean:install": "npm run clean",
+      "preversion": "npm test",
+      "protractor": "protractor",
+      "protractor:delay": "sleep 3 && npm run protractor",
+      "protractor:live": "protractor --elementExplorer",
+      "rimraf": "rimraf",
+      "server:dev:hmr": "npm run server:dev -- --hotOnly",
+      "server:aot:dev": "cross-env BUILD_AOT=1 npm run server:dev",
+      "server:dev": "npm run webpack-dev-server -- --config config/webpack.dev.js --open --progress --profile --watch --content-base src/",
+      "server:prod": "http-server dist -c-1 --cors",
+      "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
+      "server": "npm run server:dev",
+      "start:hmr": "npm run server:dev:hmr",
+      "start": "npm run server:dev",
+      "start:aot": "npm run server:aot:dev",
+      "test": "npm run lint && karma start",
+      "tslint": "tslint",
+      "typedoc": "typedoc",
+      "version": "npm run build",
+      "watch:dev:hmr": "npm run watch:dev -- --hot",
+      "watch:dev": "npm run build:dev -- --watch",
+      "watch:aot:dev": "npm run build:aot:dev -- --watch",
+      "watch:prod": "npm run build:prod -- --watch",
+      "watch:aot:prod": "npm run build:aot:prod -- --watch",
+      "watch:test": "npm run test -- --auto-watch --no-single-run",
+      "watch": "npm run watch:dev",
+      "webdriver-manager": "node ./node_modules/protractor/bin/webdriver-manager",
+      "webdriver:start": "node ./node_modules/protractor/bin/webdriver-manager start",
+      "webdriver:update": "node ./node_modules/protractor/bin/webdriver-manager update",
+      "webpack-dev-server": "node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
+      "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js"
+    },
+    "dependencies": {
+      "@angular/animations": "~5.0.1",
+      "@angular/common": "~5.0.1",
+      "@angular/compiler": "~5.0.1",
+      "@angular/core": "~5.0.1",
+      "@angular/forms": "~5.0.1",
+      "@angular/http": "~5.0.1",
+      "@angular/platform-browser": "~5.0.1",
+      "@angular/platform-browser-dynamic": "~5.0.1",
+      "@angular/platform-server": "~5.0.1",
+      "@angular/router": "~5.0.1",
+      "core-js": "^2.5.1",
+      "http-server": "^0.9.0",
+      "ie-shim": "^0.1.0",
+      "reflect-metadata": "^0.1.10",
+      "rxjs": "^5.5.2",
+      "zone.js": "~0.8.18"
+    },
+    "devDependencies": {
+      "@angular-devkit/build-optimizer": "^0.0.32",
+      "@angular/compiler-cli": "~5.0.1",
+      "@compodoc/compodoc": "^1.0.0-beta.15",
+      "@ngtools/webpack": "^1.8.0",
+      "@types/hammerjs": "^2.0.34",
+      "@types/jasmine": "2.5.45",
+      "@types/node": "^7.0.39",
+      "@types/source-map": "^0.5.0",
+      "@types/uglify-js": "^2.6.28",
+      "@types/webpack": "^2.2.16",
+      "add-asset-html-webpack-plugin": "^1.0.2",
+      "angular2-template-loader": "^0.6.2",
+      "assets-webpack-plugin": "^3.5.1",
+      "awesome-typescript-loader": "^3.3.0",
+      "codelyzer": "~2.1.1",
+      "copy-webpack-plugin": "^4.0.1",
+      "cross-env": "^5.0.0",
+      "css-loader": "^0.28.0",
+      "exports-loader": "^0.6.4",
+      "expose-loader": "^0.7.3",
+      "extract-text-webpack-plugin": "~3.0.0",
+      "file-loader": "^0.11.1",
+      "find-root": "^1.0.0",
+      "gh-pages": "^1.0.0",
+      "html-webpack-plugin": "^2.28.0",
+      "imports-loader": "^0.7.1",
+      "inline-manifest-webpack-plugin": "^3.0.1",
+      "istanbul-instrumenter-loader": "2.0.0",
+      "jasmine-core": "^2.5.2",
+      "karma": "^1.6.0",
+      "karma-chrome-launcher": "^2.0.0",
+      "karma-coverage": "^1.1.1",
+      "karma-jasmine": "^1.1.0",
+      "karma-mocha-reporter": "^2.2.3",
+      "karma-remap-coverage": "^0.1.4",
+      "karma-sourcemap-loader": "^0.3.7",
+      "karma-webpack": "^2.0.4",
+      "lint-staged": "^4.2.3",
+      "ngc-webpack": "^4.0.2",
+      "node-sass": "^4.5.2",
+      "npm-run-all": "^4.0.2",
+      "optimize-js-plugin": "0.0.4",
+      "parse5": "^3.0.2",
+      "preload-webpack-plugin": "^1.2.2",
+      "protractor": "^5.1.1",
+      "raw-loader": "0.5.1",
+      "rimraf": "~2.6.1",
+      "sass-loader": "^6.0.3",
+      "script-ext-html-webpack-plugin": "^1.8.5",
+      "source-map-loader": "^0.2.1",
+      "string-replace-loader": "~1.3.0",
+      "style-loader": "^0.18.1",
+      "to-string-loader": "^1.1.5",
+      "ts-node": "^3.3.0",
+      "tslib": "^1.7.1",
+      "tslint": "~4.5.1",
+      "tslint-loader": "^3.5.2",
+      "typedoc": "^0.7.1",
+      "typescript": "~2.5.2",
+      "uglifyjs-webpack-plugin": "^1.0.1",
+      "url-loader": "^0.5.8",
+      "webpack": "^3.8.1",
+      "webpack-dev-middleware": "^1.10.1",
+      "webpack-dev-server": "~2.7.1",
+      "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
+      "webpack-merge": "~4.1.0"
+    },
+    "repository": {
+      "type": "git",
+      "url": "https://github.com/gdi2290/angular-starter.git"
+    },
+    "bugs": {
+      "url": "https://github.com/gdi2290/angular-starter/issues"
+    },
+    "engines": {
+      "node": ">= 4.2.1",
+      "npm": ">= 3"
+    }
   }
-}
