@@ -9,7 +9,7 @@ describe('App', () => {
 
   it('should have a title', async () => {
     let subject = await browser.getTitle();
-    let result  = 'Angular2 Webpack Starter by @gdi2290 from @AngularClass';
+    let result  = 'Angular Starter by @gdi2290 from @TipeIO';
     expect(subject).toEqual(result);
   });
 
