@@ -1,32 +1,26 @@
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import {
-  NgModule,
-  ApplicationRef
-} from '@angular/core';
-import {
-  removeNgStyles,
-  createNewHosts,
-  createInputTransfer
-} from '@angularclass/hmr';
-import {
-  RouterModule,
-  PreloadAllModules
-} from '@angular/router';
+import { RouterModule, PreloadAllModules } from '@angular/router';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-
-import { ENV_PROVIDERS } from './environment';
+/*
+ * Platform and Environment providers/directives/pipes
+ */
+import { environment } from 'environments/environment';
 import { ROUTES } from './app.routes';
+// App is our top level component
 import { AppComponent } from './app.component';
 import { APP_RESOLVER_PROVIDERS } from './app.resolver';
 import { AppState, InternalStateType } from './app.service';
 import { HomeComponent } from './home';
 import { AboutComponent } from './about';
 import { NoContentComponent } from './no-content';
-import { CommonComponentsModule } from './components/common.module';
+import { XLargeDirective } from './home/x-large';
+import { DevModuleModule } from './+dev-module';
 
-import '../styles/styles.css';
+import '../styles/styles.scss';
 import '../styles/headings.css';
 
 // Application wide providers
@@ -41,85 +35,44 @@ type StoreType = {
   disposeOldHosts: () => void
 };
 
-
+/**
+ * `AppModule` is the main entry point into Angular2's bootstraping process
+ */
 @NgModule({
-  bootstrap:([
-
-      AppComponent
-
-  ]),
+  bootstrap: [ AppComponent ],
   declarations: [
     AppComponent,
     AboutComponent,
     HomeComponent,
-    NoContentComponent
+    NoContentComponent,
+    XLargeDirective
   ],
-
+  /**
+   * Import Angular's modules.
+   */
   imports: [
     BrowserModule,
+    BrowserAnimationsModule,
     FormsModule,
     HttpModule,
-    CommonComponentsModule,
-    RouterModule.forRoot(ROUTES, { useHash: true, preloadingStrategy: PreloadAllModules })
-  ],
+    RouterModule.forRoot(ROUTES, {
+      useHash: Boolean(history.pushState) === false,
+      preloadingStrategy: PreloadAllModules
+    }),
 
+    /**
+     * This section will import the `DevModuleModule` only in certain build types.
+     * When the module is not imported it will get tree shaked.
+     * This is a simple example, a big app should probably implement some logic
+     */
+    ...environment.showDevModule ? [ DevModuleModule ] : [],
+  ],
+  /**
+   * Expose our Services and Providers into Angular's dependency injection.
+   */
   providers: [
-    ENV_PROVIDERS,
+    environment.ENV_PROVIDERS,
     APP_PROVIDERS
   ]
 })
-export class AppModule {
-
-  constructor(
-    public appRef: ApplicationRef,
-    public appState: AppState
-  ) {}
-
-  public hmrOnInit(store: StoreType) {
-    if (!store || !store.state) {
-      return;
-    }
-    console.log('HMR store', JSON.stringify(store, null, 2));
-    /**
-     * Set state
-     */
-    this.appState._state = store.state;
-    /**
-     * Set input values
-     */
-    if ('restoreInputValues' in store) {
-      let restoreInputValues = store.restoreInputValues;
-      setTimeout(restoreInputValues);
-    }
-
-    this.appRef.tick();
-    delete store.state;
-    delete store.restoreInputValues;
-  }
-
-  public hmrOnDestroy(store: StoreType) {
-    const cmpLocation = this.appRef.components.map((cmp) => cmp.location.nativeElement);
-    store.state = this.appState._state;
-    /**
-     * Recreate root elements
-     */
-    store.disposeOldHosts = createNewHosts(cmpLocation);
-    /**
-     * Save input values
-     */
-    store.restoreInputValues  = createInputTransfer();
-    /**
-     * Remove styles
-     */
-    removeNgStyles();
-  }
-
-  public hmrAfterDestroy(store: StoreType) {
-    /**
-     * Display new elements
-     */
-    store.disposeOldHosts();
-    delete store.disposeOldHosts;
-  }
-
-}
+export class AppModule {}
