@@ -1,1 +1,9 @@
-import {NgModule} from "@angular/core";import {SeekBarComponent} from "./seekbar/seekbar.component";@NgModule({    exports: [SeekBarComponent],    declarations: [SeekBarComponent]})export class CommonComponentsModule {}
+import { NgModule } from "@angular/core";
+import { SeekBarComponent } from "./seekbar/seekbar.component";
+
+@NgModule({
+    exports: [SeekBarComponent],
+    declarations: [SeekBarComponent]
+})
+
+export class CommonComponentsModule {}
