@@ -1,8 +1,8 @@
-import { NgModuleRef } from "@angular/core";
+import { NgModuleRef } from '@angular/core';
 
 export interface Environment {
-    production: boolean;
-    ENV_PROVIDERS: any;
-    showDevModule: boolean;
-    decorateModuleRef(modRef: NgModuleRef<any>): NgModuleRef<any>;
+  production: boolean;
+  ENV_PROVIDERS: any;
+  showDevModule: boolean;
+  decorateModuleRef(modRef: NgModuleRef<any>): NgModuleRef<any>;
 }
