@@ -1,23 +1,29 @@
 /**
  * Angular bootstrapping
  */
-import { platformBrowser } from '@angular/platform-browser';
-import { decorateModuleRef } from './app/environment';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { environment } from 'environments/environment';
+
 /**
  * App Module
- * our top level module that holds all of our components.
+ * our top level module that holds all of our components
  */
-import { AppModuleNgFactory } from '../compiled/src/app/app.module.ngfactory';
+import { AppModule } from './app';
+
 /**
- * Bootstrap our Angular app with a top level NgModule.
+ * Bootstrap our Angular app with a top level NgModule
  */
 export function main(): Promise<any> {
-  return platformBrowser()
-    .bootstrapModuleFactory(AppModuleNgFactory)
-    .then(decorateModuleRef)
+  return platformBrowserDynamic()
+    .bootstrapModule(AppModule)
+    .then(environment.decorateModuleRef)
     .catch((err) => console.error(err));
 }
 
+/**
+ * Needed for hmr
+ * in prod this is replace for document ready
+ */
 switch (document.readyState) {
   case 'loading':
     document.addEventListener('DOMContentLoaded', _domReadyHandler, false);
@@ -29,6 +35,7 @@ switch (document.readyState) {
 }
 
 function _domReadyHandler() {
-  document.removeEventListener('DOMContentLoaded', _domReadyHandler, false);
-  main();
+ document.removeEventListener('DOMContentLoaded', _domReadyHandler, false);
+ console.log("doc ready, main....");
+ main();
 }
